chore: perform syncback
It also adds the baseplate, which is understandable.

It also adds Terrain, which is semi-understandable. There is no terrain
right now, but maybe it still wants to add that instance.

It also adds a bunch of goofy properties on Workspace even though I
have not touched them at all so they are probably defaults. This is not
understandable.